backend/src/requirements.txt
Flask==3.0.0
pyserial==3.5
bcrypt==4.1
flask-cors==4.0.0
celery==5.3.6
flask_login
opencv-python